+++ internal/search/streaming/stream_test.go
@@ -1,13 +1,16 @@
 package streaming
 
 import (
+	"encoding/json"
 	"sync"
 	"testing"
 	"time"
 
+	"github.com/hexops/autogold"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/atomic"
 
+	"github.com/sourcegraph/sourcegraph/internal/search/filter"
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
 )
 
@@ -131,4 +134,154 @@ func TestBatchingStream(t *testing.T) {
 		s.Done()
 		require.Equal(t, count.Load(), int64(10))
 	})
+}
+
+func TestWithSelect(t *testing.T) {
+	dataCopy := func() SearchEvent {
+		return SearchEvent{
+			Results: []result.Match{
+				&result.FileMatch{
+					File: result.File{Path: "pokeman/charmandar"},
+					LineMatches: []*result.LineMatch{{
+						OffsetAndLengths: make([][2]int32, 1),
+					}},
+				},
+				&result.FileMatch{
+					File: result.File{Path: "pokeman/charmandar"},
+					LineMatches: []*result.LineMatch{{
+						OffsetAndLengths: make([][2]int32, 1),
+					}},
+				},
+				&result.FileMatch{
+					File: result.File{Path: "pokeman/bulbosaur"},
+					LineMatches: []*result.LineMatch{{
+						OffsetAndLengths: make([][2]int32, 1),
+					}},
+				},
+				&result.FileMatch{
+					File: result.File{Path: "digiman/ummm"},
+					LineMatches: []*result.LineMatch{{
+						OffsetAndLengths: make([][2]int32, 1),
+					}},
+				},
+			},
+		}
+	}
+
+	test := func(selector string) string {
+		selectPath, _ := filter.SelectPathFromString(selector)
+		agg := NewAggregatingStream()
+		selectAgg := WithSelect(agg, selectPath)
+		selectAgg.Send(dataCopy())
+		s, _ := json.MarshalIndent(agg.Results, "", "  ")
+		return string(s)
+	}
+
+	autogold.Want("dedupe paths for select:file.directory", `[
+  {
+    "Path": "pokeman/",
+    "LineMatches": null,
+    "LimitHit": false
+  },
+  {
+    "Path": "digiman/",
+    "LineMatches": null,
+    "LimitHit": false
+  }
+]`).Equal(t, test("file.directory"))
+
+	autogold.Want("dedupe paths select:file", `[
+  {
+    "Path": "pokeman/charmandar",
+    "LineMatches": null,
+    "LimitHit": false
+  },
+  {
+    "Path": "pokeman/bulbosaur",
+    "LineMatches": null,
+    "LimitHit": false
+  },
+  {
+    "Path": "digiman/ummm",
+    "LineMatches": null,
+    "LimitHit": false
+  }
+]`).Equal(t, test("file"))
+
+	autogold.Want("don't dedupe file matches for select:content", `[
+  {
+    "Path": "pokeman/charmandar",
+    "LineMatches": [
+      {
+        "Preview": "",
+        "OffsetAndLengths": [
+          [
+            0,
+            0
+          ]
+        ],
+        "LineNumber": 0
+      },
+      {
+        "Preview": "",
+        "OffsetAndLengths": [
+          [
+            0,
+            0
+          ]
+        ],
+        "LineNumber": 0
+      }
+    ],
+    "LimitHit": false
+  },
+  {
+    "Path": "pokeman/charmandar",
+    "LineMatches": [
+      {
+        "Preview": "",
+        "OffsetAndLengths": [
+          [
+            0,
+            0
+          ]
+        ],
+        "LineNumber": 0
+      }
+    ],
+    "LimitHit": false
+  },
+  {
+    "Path": "pokeman/bulbosaur",
+    "LineMatches": [
+      {
+        "Preview": "",
+        "OffsetAndLengths": [
+          [
+            0,
+            0
+          ]
+        ],
+        "LineNumber": 0
+      }
+    ],
+    "LimitHit": false
+  },
+  {
+    "Path": "digiman/ummm",
+    "LineMatches": [
+      {
+        "Preview": "",
+        "OffsetAndLengths": [
+          [
+            0,
+            0
+          ]
+        ],
+        "LineNumber": 0
+      }
+    ],
+    "LimitHit": false
+  }
+]`).Equal(t, test("content"))
 }